PMM-6067 fix timerange updates
